glad_app/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
anyio==3.7.1
asn1crypto==1.5.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.4.1
autopep8==2.0.2
backoff==2.2.1
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chainlit==0.6.3
charset-normalizer==3.2.0
click==8.1.7
cloud-sql-python-connector==1.4.1
cohere==4.21
cryptography==41.0.3
dataclasses-json==0.5.14
Deprecated==1.2.14
dnspython==2.4.2
exceptiongroup==1.1.2
fastapi==0.97.0
fastapi-socketio==0.0.10
fastavro==1.8.2
filelock==3.12.2
filetype==1.2.0
Flask==2.3.3
Flask-JWT-Extended==4.5.2
frozenlist==1.4.0
fsspec==2023.6.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
google-api-core==2.16.2
google-auth==2.22.0
google-cloud==0.34.0
google-cloud-bigquery==3.17.1
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.274
langsmith==0.0.26
Lazify==0.4.0
loguru==0.7.0
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
nodeenv==1.8.0
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-exporter-otlp-proto-common==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-otlp-proto-http==1.19.0
opentelemetry-instrumentation==0.40b0
opentelemetry-proto==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
packaging==23.1
pg8000==1.30.1
pinecone-client==2.2.2
praw==7.6.0
prawcore==2.3.0
prisma==0.9.1
promptwatch==0.2.9
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==1.10.12
PyJWT==2.8.0
pyOpenSSL==23.2.0
pypdf==3.15.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.6.1
python-graphql-client==0.4.3
python-socketio==5.8.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rsa==4.9
safetensors==0.3.3
scramp==1.4.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.1
torch==2.0.1
tqdm==4.66.1
transformers==4.32.0
typing-inspect==0.9.0
typing_extensions==4.7.1
update-checker==0.18.0
uptrace==1.19.0
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
websocket-client==1.4.2
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
